fix: Using coroutines when animate theme change.
